requirements.txt
numpy
scipy
rouge_score
fire
openai
transformers
torch
langchain
black
python-dotenv
tiktoken
wikipedia
chromadb
scipy
scikit-learn
statsmodels
matplotlib
seaborn
anthropic
accelerate
bitsandbytes
sentencepiece
pydantic
vllm
